feat: allow redirection of command IO to user supplied writers
Allow the user to redirect the stdout and stderr to their own Writers.
This can be useful for advanced streaming or parsing cases.

feat: Allow overriding the shell command

When using `Shell=true` the user can now define a custom `ShellPath`.
When `ShellPath` is empty, the default `/bin/bash` will be used.

chore: Make the ExecTask a Stringer

Add the String() method so that the task can be easily printed. This
will use the same command construction logic, thus showing the shell,
command, and args that will be used.

chore: expose EnvVars method for easier debugging.

Add EnvVars() method, which allows the caller to see the envvars that
will be used when the command is executed.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>