Allow for spaces in command
Prior to this change, Windows users who had a space in the
path to a command would encounter errors.

Tested with a unit test which failed before commenting out
code in exec.go to split out the command and arguments if
a space was within the command.

This is a breaking change and downstream users should make
sure they populate Command and Args separately.

https://github.com/alexellis/arkade/issues/117

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>